+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.0.0]
+## [1.0.0] - 2023-05-17
 
 * Upgrade to Flutter 3.10 and Dart 3.0.0.
 * Updating class Decimal features.
@@ -12,6 +12,8 @@
     * Duplet
     * Triplet
 * Updating AbstractUIBuilder to use null string for labelPrefix and labelSuffix.
+* Removing labelPrefix and editController from formContent method.
+* Removing AbstractEditContent class.
 
 ## [0.22.3] - 2023-05-12